+++ intern-crm-web/src/main/java/com/smartdev/controller/manager/ManageSellerController.java
@@ -1,0 +1,15 @@
+package com.smartdev.controller.manager;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/manager_crm")
+public class ManageSellerController {
+    @RequestMapping(value="/list-seller")
+    public String listSeller(){
+        return "list-seller";
+    }
+
+
+}
