+++ intern-crm-web/src/main/webapp/WEB-INF/title3.xml
@@ -4,38 +4,48 @@
 <tiles-definitions>
 
     <!-- Base Define -->
-    <definition name="definition" template="/view/template/layout/template.jsp">
+    <definition name="definition-manager" template="/view/template/layout/template.jsp">
         <put-attribute name="header" value="/view/template/layout/header.jsp" />
-        <put-attribute name="menu" value="/view/template/layout/menu.jsp" />
+        <put-attribute name="menu" value="/view/template/layout/menu-manager.jsp" />
         <put-attribute name="body" value="" />
         <put-attribute name="footer" value="/view/template/layout/footer.jsp" />
     </definition>
 
-    <definition name="home" extends="definition">
+    <definition name="home" extends="definition-manager">
         <put-attribute name="body" value="/view/home.jsp" />
     </definition>
-    <definition name="seller-manager" extends="definition">
-        <put-attribute name="body" value="/view/seller-manager.jsp"/>
+    <definition name="list-seller" extends="definition-manager">
+        <put-attribute name="body" value="/view/list-seller.jsp"/>
     </definition>
     <definition name="login" template="/view/login.jsp">
     </definition>
 
-    <definition name="home-manager" extends="definition">
+    <definition name="home-manager" extends="definition-manager">
         <put-attribute name="body" value="/view/manager-home.jsp"/>
     </definition>
 
-    <definition name="list-custom-manage" extends="definition">
+    <definition name="list-custom-manage" extends="definition-manager">
         <put-attribute name="body" value="/view/list-custom-manage.jsp"/>
     </definition>
 
-    <definition name="list-custom-seller" extends="definition">
+
+
+
+<!-- seller-->
+    <definition name="definition-seller" template="/view/template/layout/template.jsp">
+        <put-attribute name="header" value="/view/template/layout/header.jsp" />
+        <put-attribute name="menu" value="/view/template/layout/menu-seller.jsp" />
+        <put-attribute name="body" value="" />
+        <put-attribute name="footer" value="/view/template/layout/footer.jsp" />
+    </definition>
+    <definition name="list-custom-seller" extends="definition-seller">
         <put-attribute name="body" value="/view/list-custom-seller.jsp"/>
     </definition>
-    <definition name="homeSeller" extends="definition">
-
-        <put-attribute name="body" value="/view/homeSeller.jsp" />
+    <definition name="list-custom" extends="definition-seller">
+        <put-attribute name="body" value="/view/list-custom-manage.jsp"/>
     </definition>
-    <definition name="addcustomer" extends="definition">
-            <put-attribute name="body" value="/view/addcustomer.jsp"/>
+
+    <definition name="addcustomer" extends="definition-seller">
+        <put-attribute name="body" value="/view/addcustomer.jsp"/>
     </definition>
 </tiles-definitions>
