--- intern-crm-web/src/main/webapp/view/homeSeller.jsp
@@ -1,60 +1,0 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<div>
-    <div class="input-group mt-5" style="width: 30%">
-        <div class="input-group-prepend">
-            <label class="input-group-text" for="inputGroupSelect01">Status</label>
-        </div>
-        <select class="custom-select" id="inputGroupSelect01">
-            <option selected>Choose...</option>
-            <option value="">Prospect</option>
-            <option value="">Lead</option>
-            <option value="">Potential Lead</option>
-            <option value="">Active Lead</option>
-        </select>
-    </div>
-
-    <div>
-        <h3 align="center">THÔNG TIN KHÁCH HÀNG</h3>
-        <table class="table table-hover">
-            <thead>
-            <tr>
-
-                <th>Id</th>
-                <th>Name</th>
-                <th>Birthday</th>
-                <th>Phone</th>
-                <th>Email</th>
-                <th>Fabebook</th>
-                <th>Status</th>
-
-            </tr>
-            </thead>
-            <tbody>
-            <tr>
-                <td>1</td>
-                <td>NgocAnh</td>
-
-                <td>1/1/1997</td>
-                <td>016987525565</td>
-                <td>abc@gmail.com</td>
-                <td>FHDJ</td>
-                <td>LEAD
-
-                </td>
-
-            </tr>
-            <tr>
-                <td>1</td>
-                <td>NgocAnh</td>
-
-                <td>1/1/1997</td>
-                <td>016987525565</td>
-                <td>abc@gmail.com</td>
-                <td>FHDJ</td>
-                <td>LEAD</td>
-            </tr>
-            </tbody>
-        </table>
-
-    </div>
-</div>
