+++ intern-crm-web/src/main/webapp/view/template/layout/menu-manager.jsp
@@ -9,10 +9,13 @@
     <div class="sidebar bg-menu">
     <ul class="list-unstyled my-0">
     <li class="p-3 active ">
-    <a href="/manage-seller" class="text-white lead ">Seller Manager</a>
+    <a href="/manager_crm/home" class="text-white lead ">Home</a>
+    </li>
+    <li class="p-3  ">
+    <a href="/manager_crm/list-seller" class="text-white lead ">Seller Manager</a>
     </li>
     <li class="p-3">
-    <a href="/list-custom-manage" class="text-white lead">Customer Manager</a>
+    <a href="/manager_crm/list-custom" class="text-white lead">Customer Manager</a>
     </li>
     <li class="p-3">
     <a href="#" class="text-white lead ">Block list</a>
