+++ intern-crm-web/src/main/webapp/view/template/layout/menu-seller.jsp
@@ -1,0 +1,21 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Ha Nguyen
+  Date: 6/14/2018
+  Time: 3:49 PM
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+    <div class="sidebar bg-menu">
+    <ul class="list-unstyled my-0">
+    <li class="p-3 active ">
+    <a href="/seller/list-custom-seller" class="text-white lead ">Home</a>
+    </li>
+    <li class="p-3">
+    <a href="/seller/list-custom" class="text-white lead">Customer Manager</a>
+    </li>
+    <li class="p-3">
+    <a href="#" class="text-white lead ">Block list</a>
+    </li>
+    </ul>
+    </div>
